% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_hourly.R
\name{daily_to_hourly}
\alias{daily_to_hourly}
\title{Interpolate hourly temperatures from daily max/min temperatures}
\usage{
daily_to_hourly(df, add_location_key = FALSE, keep_suntimes = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} of daily max/min temperature observations in
degrees celsius and geographic coordinates in decimal degree format.
The \code{data.frame} must have, at a minimum, five variables named "date"
(Date), "lat" (double), "lon" (double), "min" (double), and "max" (double).
All columns in the input \code{data.frame} are preserved in the output.}

\item{add_location_key}{Optionally generate a unique location key (character
string) to the output \code{data.frame} for each set of unique geographic
coordinates, if one does not already exist. This is useful when modelling
development at multiple locations using the function (ADD FUNCTION NAME).}

\item{keep_suntimes}{Whether to keep columns with the sunrise and sunset times in the output}
}
\value{
A \code{tibble} # we should default to \code{data.frame}
}
\description{
Interpolates hourly temperature observations from daily maximum and minimum
temperatures for a given geographic location, using the
model of Cesaraccio (2001).
}
\examples{
# Single location
df1  <- example_df(nlocations = 1, ndays = 90, start_date = "2024-03-01")
head(df1)
h1 <- daily_to_hourly(df1)
head(h1)
tail(h1)
# PLOT THE DATA - ADD CODE

# Multiple locations
df3  <- example_df(nlocations = 3, ndays = 90, start_date = "2024-03-01")
head(df3)
tail(df3)
dim(df3)
h3 <- d2h(df3)
head(h3)
tail(h3)
dim(h3)
# PLOT THE DATA - ADD CODE

}
\references{
Cesaraccio et al. (2001). An improved model for determining
degree-day values from daily temperature data. International Journal of
Biometeorology 45.4 (2001): 161-169.
}
