% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmdev-package.R
\docType{package}
\name{dbmdev}
\alias{dbmdev}
\alias{dbmdev-package}
\title{dbmdev: Predict incremental temperature-based development of diamondback moth}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kym Perry \email{kym.perry@sa.gov.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Stephen M. Pederson [contributor]
}

}
\keyword{internal}
