% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_params.R
\name{dev_params}
\alias{dev_params}
\title{Stage-specific development parameters for the diamondback moth for Briere's model 2}
\usage{
dev_params(species = "dbm")
}
\arguments{
\item{species}{Diamondback moth (species = "dbm")}
}
\value{
a Matrix
}
\description{
Stage-specific development parameters for the diamondback moth for Briere's model 2
}
\examples{
# Parameters for diamondback moth
dev_params()

}
