% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly.R
\name{.hourly_obs}
\alias{.hourly_obs}
\title{function to apply to a single df row with the model params}
\usage{
.hourly_obs(df)
}
\description{
function to apply to a single df row with the model params
}
\keyword{internal}
