% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_df.R
\name{example_df}
\alias{example_df}
\title{Create an example dataframe with daily max/min temperature data
for testing modelling functions}
\usage{
example_df(nlocations = 1, ndays, start_date, seed = NULL)
}
\arguments{
\item{nlocations}{The number of locations to include}

\item{ndays}{The number of days}

\item{start_date}{The starting date in YYYY-MM-DD format (character)}

\item{seed}{A random number that (if set) enables reproducible data.frames}
}
\value{
A \code{data.frame}
}
\description{
Create an example dataframe with daily max/min temperature data
for testing modelling functions
}
\examples{
# Single location
df1 <- example_df(nlocations = 1, ndays = 90, start_date = "2024-03-01")
head(df1)

# several locations
df3 <- example_df(nlocations = 3, ndays = 90, start_date = "2024-03-01")
head(df3)
tail(df3)
}
