% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBOMstdata.R
\name{formatBOMstdata}
\alias{formatBOMstdata}
\title{Format a climate dataframe for analysis}
\usage{
formatBOMstdata(df, useDST = TRUE, timeInt = 60)
}
\arguments{
\item{df}{A \code{data.frame} containing half hourly temperature observations
rom weather stations, in the format provided by the Australian Bureau of Meterology
climate data online service.}

\item{timeInt}{The time interval of data observation in minutes. Default value is
60 (Currently, this function will only aggregate data to this time interval).}
}
\value{
A reformatted \code{data.frame} with three columns, `station` (numeric
weather station code), `datetime` in POSIXct format, and `obs` (hourly temperature
observations).
}
\description{
Takes a \code{data.frame} containing half hourly temperature
observations in the format provided by the Australian Bureau of Meterology
climate data online service. It formats columns and aggregates data to hourly
observations.
}
