% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly.R
\name{hourly}
\alias{hourly}
\title{Interpolate hourly temperatures from daily max/min temperatures}
\usage{
hourly(df, add_location_key = FALSE, keep_suntimes = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} of daily maximum and minimum temperature
observations in degrees celsius and geographic coordinates in decimal
degree format. The \code{data.frame} must have, at a minimum, five variables
named "date" (Date), "lat" (double), "lon" (double), "min" (double), and
"max" (double). All columns in the input \code{data.frame} are preserved in
the output.}

\item{add_location_key}{Optionally generate a variable to the output
\code{data.frame} with a unique location key (character string) for each set
of unique geographic coordinates, if one does not already exist. This is
useful when modelling development at multiple locations using the function
\code{predict_development}.}

\item{keep_suntimes}{Whether to preserve the sunrise and sunset
times used for calculations in the output.}
}
\value{
\code{data.frame}
}
\description{
\code{hourly} interpolates hourly temperature observations from daily maximum
and minimum temperatures for a given geographic location using the
model of Cesaraccio (2001). Hourly temperatures can be used to predict
insect development using \code{predict_development}.
}
\examples{
library(dplyr)
library(ggplot2)

# Single location with 3 days
daily1  <- sample_df()
head(daily1)

hourly1 <- hourly(daily1)
head(hourly1)
tail(hourly1)
# PLOT THE DATA - ADD CODE

# 3 locations with 5 days data from a specified date
daily3  <- sample_df(locations = 3, days = 5, start_date = "2024-03-01")
print(daily3)

hourly3 <- hourly(daily3)
head(hourly3)
tail(hourly3)
dim(hourly3)

# Plot temperatures
hourly3 \%>\%
  ggplot(aes(x = datetime, y = obs)) +
  geom_line() +
  geom_point() +
  facet_wrap(~location_key) +
  theme_bw() +
  theme(aspect.ratio = 1) +
  labs(x = NULL, y = "Temperature (oC)")
}
\references{
Cesaraccio et al. (2001). An improved model for determining
degree-day values from daily temperature data. International Journal of
Biometeorology 45.4 (2001): 161-169.
}
