% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourlyTemperatures.R
\name{hourlyTemperatures}
\alias{hourlyTemperatures}
\title{Interpolate hourly temperatures from daily temperature data}
\usage{
hourlyTemperatures(df)
}
\arguments{
\item{df}{A \code{data.frame} containing columns called `station` (a unique character
string or numeric location or weather station identifier), `date` in \code{as.Date}
or \code{as.POSIXct} format, `min` and `max` (numeric daily temperatures), `sunrise` and `sunset`
(datetimes in POSIXct format).}
}
\value{
A dataframe with the variables `station`, `datetime` in as.POSIXct format, and
the interpolated hourly `obs`.
}
\description{
Implements Cesaraccio's model (see refs) for interpolating hourly
temperatures from daily maximum and minimum temperatures.
}
\details{
Uses two sine functions and a square root function to estimate the hourly
temperatures for three periods of each day. Sunrise and sunset times required
for the model can be calculated using \code{\link{DBMdevmod2::msunriset()}}
which conveniently handles \code{data.frames} with multiple dates and locations.
}
\examples{
library(dplyr)
library(ggplot2)

# create a dataframe with daily max and minimum temperatures for one week
# include latitude, longitude here for calculation of required sunrise/set times
dates <- seq(as.Date("2016-03-01"), as.Date("2016-03-08"), by = "day")
df <- data_frame(station = "loc1",
                 latitude = -31.141980,
                 longitude = 138.395350,
                 date = dates,
                 min = rnorm(n = length(dates), mean = 12, sd = 1),
                 max = min + runif(1, min = 6, max = 10)) \%>\%
              DBMdevmod2::msunriset()

# estimate hourly temperatures
# note, `obs` for the last date in the sequence will contain a string of NAs
# because the model uses minimum temperatures from the next day
hourly <- hourlyTemperatures(df)
ggplot(hourly \%>\% filter(!is.na(obs)), aes(x = datetime, y = obs)) +
  geom_line()

}
\references{
Cesaraccio, Carla, et al. "An improved model for determining degree-day values from
daily temperature data." International journal of biometeorology 45.4 (2001): 161-169.
}
