% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revDev.R
\name{revDev}
\alias{revDev}
\title{Predict reverse temperature-based development of diamondback moth
over multiple lifestages and generations based on supplied parameters}
\usage{
revDev(tempObsDf, devParamsDf, startDate, startTime = 12, startStage,
  startDev, gens, output = "increments", timeInt = 60, ...)
}
\arguments{
\item{tempObsDf}{A dataframe containing temperature observations at specified time intervals
in a variable named `obs`, and POSIXct datetimes in a variable named `datetime`.}

\item{devParamsDf}{A dataframe containing the four insect development parameters required for
Briere's model. The parameters `a`, `Tmin`, `Tmax` and `m` are held as separate variables with
the parameters for each lifestage on separate rows, with row names as character strings
corresponding to each lifestage (when modelling, the user input for `startStage` must match these).}

\item{startDate}{The starting date for the model as a character string in "yyyy-mm-dd" format.}

\item{startTime}{A number between 0 and 23 specifying the hour of the day to start the model.
The default value is 12 (midday). The model has not yet been tested using other startTimes.}

\item{startStage}{The starting lifestage as a character string. Must match one of
\code{rownames(devParamsDf)}.}

\item{startDev}{The starting cumulative development point for `startStage` as a proportion
from 0 to 1.}

\item{gens}{The number of consecutive generations to run the model.}

\item{output}{A character string specifying whether to output all incremental development predictions
("increments", the default value) or predictions summarised for each stage ("stages").}

\item{timeInt}{The time interval of increments in minutes. Th default value = 12.
The model has not yet been tested with other time intervals.}

\item{...}{Other arguments to pass to the function.}
}
\value{
A list. Each element will be a subset of the original \code{data.frame} with
the added columns `dev`, `totalDev`, `stage` and `gen`, and containing only the rows/timepoints
for a single generation. To combine output into a single \code{\link{dataframe}}, use \code{\link{dplyr::bind_rows}}.
}
\description{
Predicts reverse-in-time temperature-dependent development for multiple
consecutive DBM lifestages and generations, based on supplied parameters. A separate
analagous function, \code{\link{fwdDev}}, predicts reverse-in-time development.
}
\details{
Implements Briere's temperature-dependent insect development model,
equation 2 (Briere 1999) reverse-in-time. Outputs DBM development for each time
increment as a proportion from `0` (no development) to `1` (completed development) for
individual lifestages consecutively over multiple lifestages and generations.
}
\note{
The function has not yet been tested with `startTime` other than 12
or `timeInt` other than 60. Recommend not to vary these until tested.
}
\examples{
library(lubridate)

# create example dataframe with hourly temperature obs
date <- sort(rep(seq(as.Date("2014-06-01"), (as.Date("2014-09-08") + 100), by = "day"), times = 24))
hour <- rep(0:23, time = length(date)/24)
datetime <- ymd_hms(paste(date, hour), truncated = 2)
obs <- runif(n = length(date), min = 0, max = 38)
df <- data.frame(datetime, obs)

devParams <- data.frame(st1 = c(0.0003592, 1.754, 35.08, 4.347),
                     st2 = c(0.00036, -2.9122, 32.4565, 60.5092),
                     st3 = c(0.0003592, 1.754, 35.08, 4.347))
 rownames(devParams) <- c("a", "Tmin", "Tmax", "m")
 devParams <- t.data.frame(devParams)

# Back-predict development over 2 generations
rdev <- revDev(tempObsDf = df, devParams, startDate = "2014-09-01", startStage = "st3", startDev = 0.5, gens = 2)
rdev

# With stages summary output
rstages <- revDev(tempObsDf = df, devParams, startDate = "2014-09-01", startStage = "st3", startDev = 0.5, gens = 2, output = "stages")

# With generations summary output
rgens <- revDev(tempObsDf = df, devParams, startDate = "2014-09-01", startStage = "st3", startDev = 0.5, gens = 2, output = "generations")


}
\references{
Brazzale, A.R. (2005). hoa: An R package bundle for higher order
likelihood inference. Rnews, 5/1 May 2005, 20-27. ISSN 609-3631,
URL ftp://cran.r-project.org/doc/Rnews/Rnews_2005-1.pdf.

Briere, Jean-Francois, et al. (1999) A novel rate model of
temperature-dependent development for arthropods. Environmental Entomology 28.1: 22-29.
}
