% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dev.R
\name{plot_dev}
\alias{plot_dev}
\title{Plot development predictions for the diamondback moth}
\usage{
plot_dev(df)
}
\arguments{
\item{df}{A \code{data.frame} with developmental predictions output by
\code{predict_dev}.}

\item{what}{What output to plot, either "increments", "stages",
or "gens" (character).}

\item{display}{Optionally display the plot in the console, TRUE or
FALSE (logical).}
}
\value{
A \code{ggplot} object
}
\description{
Plot development predictions for the diamondback moth
}
\examples{
library(dplyr)
library(ggplot2)
library(RcolorBrewer)

## Note: Works properly for 1 gen only. Draft function.

# Generate some sample temperatures at a given location
d1 <- daily(days = 200, start_date = "2023-09-01")
h1 <- hourly(d1)

# Predict forward 2 generations from the "egg" stage
out1 <- predict_dev(h1, start_date = "2023-09-02", gens = 1)

# Plot
plot_dev(out1)

## End
}
