% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dev.R
\name{plot_dev}
\alias{plot_dev}
\title{Plot development predictions for the diamondback moth}
\usage{
plot_dev(
  x,
  what = "stages",
  ...,
  xlab = NULL,
  ylab = NULL,
  date_fmt = "\%d-\%b"
)
}
\arguments{
\item{x}{Developmental predictions as output by \code{predict_dev}.}

\item{what}{What output to plot, either "increments", "stages",
or "gens" (character).}

\item{...}{Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}}

\item{xlab, ylab}{Axis labels}

\item{date_fmt}{Format for printing x-axis labels. See \code{\link[base:strptime]{base::strptime()}}}
}
\value{
A \code{ggplot} object
}
\description{
Plot development predictions for the diamondback moth
}
\examples{
library(dplyr)
library(RColorBrewer)

## Note: Works properly for 1 gen only. Draft function.

# Generate some sample temperatures at a given location
d1 <- daily(days = 200, start_date = "2023-09-01")
h1 <- hourly(d1)

# Predict forward 2 generations from the "egg" stage
out1 <- predict_dev(h1, start_date = "2023-09-02", gens = 1)

# Plot
plot_dev(out1)

## End
}
